+++ flutter-hms-location/android/src/main/java/com/huawei/hms/flutter/location/handlers/FusedLocationMethodHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -72,11 +72,11 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
 
     private final Map<Integer, PendingIntent> requests;
 
-    private final SettingsClient settingsClient;
+    private SettingsClient settingsClient;
 
-    private final FusedLocationProviderClient service;
+    private FusedLocationProviderClient service;
 
-    private final LocationEnhanceService enhanceService;
+    private LocationEnhanceService enhanceService;
 
     private LogConfig logConfig;
 
@@ -89,84 +89,137 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         this.channel = channel;
         callbacks = new HashMap<>();
         requests = new HashMap<>();
+    }
+
+    private void initFusedLocationService(final MethodChannel.Result result) {
         settingsClient = LocationServices.getSettingsClient(activity);
         service = LocationServices.getFusedLocationProviderClient(activity);
         enhanceService = LocationServices.getLocationEnhanceService(activity);
+        Log.i(TAG, "Fused Location Service has been initialized.");
+        result.success(null);
     }
 
     private void checkLocationSettings(final MethodCall call, final MethodChannel.Result result) {
         final LocationSettingsRequest request = LocationUtils.fromMapToLocationSettingsRequest(call.arguments());
         this.result = result;
 
+        if (settingsClient == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         settingsClient.checkLocationSettings(request)
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new LocationSettingsFailureListener(result, activity));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new LocationSettingsFailureListener(result, activity));
+
     }
 
     private void getLastLocation(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.getLastLocation()
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
     }
 
     private void getLastLocationWithAddress(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.getLastLocationWithAddress(
                 LocationUtils.fromMapToLocationRequest(call.<Map<String, Object>>arguments()))
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
     }
 
     private void getLocationAvailability(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.getLocationAvailability()
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
     }
 
     private void setMockMode(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.setMockMode(call.<Boolean>arguments())
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
     }
 
     private void setMockLocation(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.setMockLocation(LocationUtils.fromMapToLocation(call.arguments()))
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
     }
 
     private void requestLocationUpdates(final MethodCall call, final MethodChannel.Result result) {
         final Pair<Integer, PendingIntent> intentData = buildLocationIntent();
         final LocationRequest request = LocationUtils.fromMapToLocationRequest(call.<Map<String, Object>>arguments());
 
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.requestLocationUpdates(request, intentData.second)
-                .addOnSuccessListener(
-                        new RequestUpdatesSuccessListener(call.method, activity, result, intentData.first))
-                .addOnFailureListener(
-                        new RequestUpdatesFailureListener<>(call.method, activity, result, intentData.first, requests));
+            .addOnSuccessListener(new RequestUpdatesSuccessListener(call.method, activity, result, intentData.first))
+            .addOnFailureListener(
+                new RequestUpdatesFailureListener<>(call.method, activity, result, intentData.first, requests));
+
     }
 
     private void requestLocationUpdatesCb(final MethodCall call, final MethodChannel.Result result) {
         final LocationRequest request = LocationUtils.fromMapToLocationRequest(call.<Map<String, Object>>arguments());
         final Pair<Integer, LocationCallbackHandler> callbackData = buildCallback(call.method);
 
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.requestLocationUpdates(request, callbackData.second, Looper.getMainLooper())
-                .addOnSuccessListener(
-                        new RequestUpdatesSuccessListener(call.method, activity, result, callbackData.first))
-                .addOnFailureListener(
-                        new RequestUpdatesFailureListener<>(call.method, activity, result, callbackData.first,
-                                callbacks));
+            .addOnSuccessListener(new RequestUpdatesSuccessListener(call.method, activity, result, callbackData.first))
+            .addOnFailureListener(
+                new RequestUpdatesFailureListener<>(call.method, activity, result, callbackData.first, callbacks));
+
     }
 
     private void requestLocationUpdatesExCb(final MethodCall call, final MethodChannel.Result result) {
         final LocationRequest request = LocationUtils.fromMapToLocationRequest(call.<Map<String, Object>>arguments());
         final Pair<Integer, LocationCallbackHandler> callbackData = buildCallback(call.method);
 
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.requestLocationUpdatesEx(request, callbackData.second, Looper.getMainLooper())
-                .addOnSuccessListener(
-                        new RequestUpdatesSuccessListener(call.method, activity, result, callbackData.first))
-                .addOnFailureListener(
-                        new RequestUpdatesFailureListener<>(call.method, activity, result, callbackData.first,
-                                callbacks));
+            .addOnSuccessListener(new RequestUpdatesSuccessListener(call.method, activity, result, callbackData.first))
+            .addOnFailureListener(
+                new RequestUpdatesFailureListener<>(call.method, activity, result, callbackData.first, callbacks));
+
     }
 
     private void removeLocationUpdates(final MethodCall call, final MethodChannel.Result result) {
@@ -175,11 +228,16 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         if (!requests.containsKey(incomingRequestCode)) {
             result.error(Error.NON_EXISTING_REQUEST_ID.name(), Error.NON_EXISTING_REQUEST_ID.message(), null);
         } else {
+            if (service == null) {
+                result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+                return;
+            }
+
             service.removeLocationUpdates(requests.get(incomingRequestCode))
-                    .addOnSuccessListener(
-                            new RemoveUpdatesSuccessListener<>(call.method, activity, result, incomingRequestCode,
-                                    requests))
-                    .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+                .addOnSuccessListener(
+                    new RemoveUpdatesSuccessListener<>(call.method, activity, result, incomingRequestCode, requests))
+                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
         }
     }
 
@@ -189,24 +247,34 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         if (!callbacks.containsKey(callbackId)) {
             result.error(Error.NON_EXISTING_REQUEST_ID.name(), Error.NON_EXISTING_REQUEST_ID.message(), null);
         } else {
+            if (service == null) {
+                result.error("-1", "FusedLocationService is not initialized.", null);
+                return;
+            }
+
             service.removeLocationUpdates(callbacks.get(callbackId))
-                    .addOnSuccessListener(
-                            new RemoveUpdatesSuccessListener<>(call.method, activity, result, callbackId, callbacks))
-                    .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+                .addOnSuccessListener(
+                    new RemoveUpdatesSuccessListener<>(call.method, activity, result, callbackId, callbacks))
+                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
         }
     }
 
     private void getNavigationContextState(final MethodCall call, final MethodChannel.Result result) {
         final NavigationRequest request = LocationUtils.fromMapToNavigationRequest(call.arguments());
-
+        if (enhanceService == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
         enhanceService.getNavigationState(request)
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+
     }
 
     private void enableBackgroundLocation(final MethodCall call, final MethodChannel.Result result) {
         int hasPermission = checkSelfPermission(activity.getApplicationContext(),
-                android.Manifest.permission.FOREGROUND_SERVICE);
+            android.Manifest.permission.FOREGROUND_SERVICE);
         if (hasPermission == -1) {
             result.error("NO_PERMISSION", "App does not have FOREGROUND_SERVICE permission.", null);
         }
@@ -219,7 +287,7 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         int notificationId = call.argument("notificationId");
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             NotificationManager notificationManager = (NotificationManager) activity.getApplicationContext()
-                    .getSystemService(Context.NOTIFICATION_SERVICE);
+                .getSystemService(Context.NOTIFICATION_SERVICE);
             String channelId = "com.huawei.hms.location.flutter.LOCATION_NOTIFICATION";
             NotificationChannel notificationChannel = new NotificationChannel(channelId, channelName, priority);
             notificationManager.createNotificationChannel(notificationChannel);
@@ -235,28 +303,44 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
             mNotification = builder.getNotification();
         }
 
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.enableBackgroundLocation(notificationId, mNotification)
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(e -> Log.e(TAG, e.getMessage()));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(e -> Log.e(TAG, e.getMessage()));
 
     }
 
     private void disableBackgroundLocation(final MethodCall call, final MethodChannel.Result result) {
+        if (service == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         service.disableBackgroundLocation()
-                .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
-                .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
+            .addOnSuccessListener(new DefaultSuccessListener<>(call.method, activity, result))
+            .addOnFailureListener(new DefaultFailureListener(call.method, activity, result));
     }
 
     private void setLogConfig(final MethodCall call, final MethodChannel.Result result) {
         int readStorage = checkSelfPermission(activity.getApplicationContext(),
-                android.Manifest.permission.READ_EXTERNAL_STORAGE);
+            android.Manifest.permission.READ_EXTERNAL_STORAGE);
         int writeStorage = checkSelfPermission(activity.getApplicationContext(),
-                android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
+            android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
         if (readStorage == -1 && writeStorage == -1) {
             result.error("NO_PERMISSION", "App does not have storage permission.", null);
         }
 
         logConfig = LocationUtils.fromMapToLogConfig(call.arguments());
+
+        if (settingsClient == null) {
+            result.error("-1", Error.FUSED_LOCATION_NOT_INITIALIZED.message(), null);
+            return;
+        }
+
         settingsClient.setLogConfig(logConfig).addOnFailureListener(e -> Log.e(TAG, e.getMessage()));
         if (isLogFilePath(logConfig.getLogPath())) {
             result.success("success");
@@ -278,7 +362,7 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
 
     private Pair<Integer, LocationCallbackHandler> buildCallback(final String methodName) {
         final LocationCallbackHandler callBack = new LocationCallbackHandler(activity.getApplicationContext(),
-                methodName, ++requestCode, channel);
+            methodName, ++requestCode, channel);
         callbacks.put(requestCode, callBack);
         return Pair.create(requestCode, callBack);
     }
@@ -289,11 +373,11 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         intent.setAction(Action.PROCESS_LOCATION);
         final PendingIntent pendingIntent;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            pendingIntent = PendingIntent.getBroadcast(activity.getApplicationContext(), ++requestCode,
-                    intent, PendingIntent.FLAG_UPDATE_CURRENT|PendingIntent.FLAG_MUTABLE);
+            pendingIntent = PendingIntent.getBroadcast(activity.getApplicationContext(), ++requestCode, intent,
+                PendingIntent.FLAG_UPDATE_CURRENT | PendingIntent.FLAG_MUTABLE);
         } else {
-            pendingIntent = PendingIntent.getBroadcast(activity.getApplicationContext(), ++requestCode,
-                    intent, PendingIntent.FLAG_UPDATE_CURRENT);
+            pendingIntent = PendingIntent.getBroadcast(activity.getApplicationContext(), ++requestCode, intent,
+                PendingIntent.FLAG_UPDATE_CURRENT);
         }
         requests.put(requestCode, pendingIntent);
         return Pair.create(requestCode, pendingIntent);
@@ -304,6 +388,9 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
         HMSLogger.getInstance(activity.getApplicationContext()).startMethodExecutionTimer(call.method);
 
         switch (call.method) {
+            case "initFusedLocationService":
+                initFusedLocationService(result);
+                break;
             case "checkLocationSettings":
                 checkLocationSettings(call, result);
                 break;
@@ -367,13 +454,13 @@ public class FusedLocationMethodHandler implements MethodChannel.MethodCallHandl
             if (resultCode == Activity.RESULT_OK) {
                 final LocationSettingsStates states = LocationSettingsStates.fromIntent(intent);
                 HMSLogger.getInstance(activity.getApplicationContext())
-                        .sendSingleEvent("checkLocationSettings.onActivityResult");
+                    .sendSingleEvent("checkLocationSettings.onActivityResult");
                 incomingResult.success(LocationUtils.fromLocationSettingsStatesToMap(states));
             } else {
                 HMSLogger.getInstance(activity.getApplicationContext())
-                        .sendSingleEvent("checkLocationSettings" + ".onActivityResult", "-1");
+                    .sendSingleEvent("checkLocationSettings" + ".onActivityResult", "-1");
                 incomingResult.error(Error.LOCATION_SETTINGS_NOT_AVAILABLE.name(),
-                        Error.LOCATION_SETTINGS_NOT_AVAILABLE.message(), null);
+                    Error.LOCATION_SETTINGS_NOT_AVAILABLE.message(), null);
             }
         }
